requirements.txt
fastapi==0.68.0
uvicorn==0.15.0
Pillow==8.3.2
torch==1.9.0
torchvision==0.10.0
matplotlib==3.4.3
numpy==1.21.2
tqdm==4.62.3
